chore: refactor the refactoring
